gsttemplatematch.h: Explicitly use OpenCV's enum values
The `stbt-templatematch` GStreamer element's `matchMethod` property
takes values identical to the OpenCV constants taken by the OpenCV
`cvMatchTemplate` function.

We do need our own enum (`GstTMMatchMethod`) because it is required by
GStreamer's GObject boilerplate. But now we can have the compiler
enforce that the values are kept in sync with the OpenCV equivalents,
instead of listing the OpenCV constants in a comment that the compiler
can't check.

This change is also necessary for the next commit.