gsttemplatematch.[ch]: Don't allow nonsensical `matchMethod` values
Instead of printing a warning that essentially says "this isn't going to
work", fail hard at pipeline parse time. GStreamer output is very
verbose, so warnings are all too easily missed.

The way the templatematch algorithm works is:

1. Call OpenCV's `cvMatchTemplate` function which returns a "heat map"
   where the brightest pixels indicate the most likely matches.

   http://docs.opencv.org/modules/imgproc/doc/object_detection.html#matchtemplate

2. Select the brightest point in the heat map.

3. Check that it's above the threshold specified by `matchThreshold`.

The `cvMatchTemplate` function in (1) takes a parameter specifying which
algorithm to use; our stbt-templatematch GStreamer element passes in the
value of its `matchMethod` property. There are 3 algorithms (the
differences are not interesting for this discussion), and each of them
comes in 2 varieties: Raw and Normalised.

The "normalised" output (i.e. the values in each pixel in the heat map)
will be between 1 and 0, where closer to 1 means more likely to be a
real match. The values from the "raw" version are arbitrary numbers;
their interpretation depends on the size of the input images.

Thus, it is impossible to specify a `matchThreshold` that is meaningful
for any arbitrary template image, unless we use the "normalised" image.
Even if you're working with a single known template image, there isn't
anything you can do with the "raw" versions that you can't also achieve
with the "normalised" equivalent.