Revert to using the original templatematch algorithm by default
Commit 06f34140 did two things:

1. Add a new algorithm ("normed-absdiff") for the second pass (the
   "confirm" pass) of the templatematch algorithm.

2. Made the new algorithm the default.

This commit reverts the changes in 06f34140 that made the new algorithm
the default, and also reverts the changes to selftests in that same
commit that show why making it the default is a bad idea.

It alarms me that 06f34140 had to use the old algorithm to avoid false
matches with the template image "videotestsrc-bw-flipped.png", which is
as different as it could possibly be from the source video. However this
seems to be a false alarm -- I can't actually reproduce a false match
of this template image.

However, throughout real-world testing, the new algorithm has
failed (giving a false match or false nomatch) in some situations where
the original algorithm was working. So to avoid disruption to users,
we'll use the old algorithm by default; users can switch to the new
algorithm for specific cases by setting
`confirm_method="normed-absdiff"` in calls to `wait_for_match`, or
globally by setting `confirm_method=normed-absdiff` in stbt.conf.