README: Add Travis CI build status icon
So that the build status will be visible in the GitHub web interface
at https://github.com/drothlis/stb-tester (which displays the README
in the main project page).

travis-ci.org is a free Continuous Integration service -- see commit
b08660e5 for details.

When we generate the stbt man page from this README, we have to remove
this image, because man pages can't have images and we'd end up with
an ugly placeholder instead.