Configuration in tool-specific sections no longer overrides "global" section
This affects the loading of default values from the configuration file
for "stbt run" and "stbt record", and the behaviour of the python
function "stbt.get_config" and of the command-line utility
"stbt config".

Note that configuration items that are only used by a single tool still
remain in the tool-specific section. Those items now must be in the
tool-specific section; they won't be searched for in the "global"
section.

UPGRADING
=========

This is a non-backwards-compatible change. If you had customised your
config file, you will need to ensure it complies with the new expected
layout:

* Move these entries from the "[run]" or "[record]" sections to the
  "[global]" section:

  - control
  - source_pipeline
  - sink_pipeline
  - verbose

* This entry must be in the "[run]" section:

  - script

* These entries must be in the "[record]" section:

  - output_file
  - control_recorder

BACKGROUND
==========

The tools like "stbt run" and "stbt record" take parameters on the
command line, like "--source-pipeline". If not specified on the command
line, the value is taken from the configuration file(s). In a typical
installation, the following files are searched:

* /etc/stbt/stbt.conf
* ~/.config/stbt/stbt.conf
* $STBT_CONFIG_FILE

(with later files taking precedence).

These files use the "ini" format:

    [section]
    key = value

Prior to this commit, "stbt run" searches for each configuration item
(e.g. "source_pipeline") in the "[run]" section; if not found, then it
searches in the "[global]" section.

The confusing behaviour was that "[run] source_pipeline" specified in
/etc/stbt/stbt.conf would override "[global] source_pipeline" specified
in ~/.config/stbt/stbt.conf (note that if both items were specified in
the same section, the precedence would be the opposite order).

Nobody actually uses this "feature" -- there hasn't been a need for a
particular tool to override the global configuration used by all other
tools.