stbt-record: check pipeline is delivering frames before any "recording"
*~ TEMPORY COMMIT ~*

Previously, if a user attempted to record a test when no frame buffers
were being received, upon receiving a second† control press signal, the
program would fatal error because there is no frame to "save" as a
screenshot.

†Second press because stbt-record goes on the assumption that a user will
only give a key press once the action as a result of the previous key press
has completed. Ergo, the process of recording a test goes as such:

1) stbt-record initiates, prompts user to start pressing keys
2) User gives key press event
3) stbt receiver and passes on key press event
3) User waits until actions complete from key press
4) User gives second key press event
5) stbt receives the key press event
6) stbt takes a saves a screenshot
7) stbt passes on the key press event
8) User waits until actions complete from key press
9) etc...

Although stbt-record (tries) to receive the current frame buffer after the
first button press, it doesn't save the frame, which is what causes the
fatal error.

What this commit does is, firstly, check whether frames are being delivered
and secondly, if it is not getting getting frames, it shuts down the
pipeline and reinitiates it. It attempts this 5 times, each time waiting
for 5 seconds before restarting the pipeline after it closes it, and then
exits if frames still aren't being delivered.