stbt-record.record: won't try to save frames from empty frame buffers
This is an additional precaution to the previous commit, however whereas
the previous commit won't allow user input until the pipeline has started,
and therefore captured frames can be saves to disk, this commit prevents
key press events being sent and (attempted) screenshots taken if there are
no frames on the framebuffer because video has been lost.

Attempting to take a screenshot would cause stbt-record to fatal error,
but not before the key press event had been sent, which could (probably
will) change the state of the device-under-test, and so will have to be
re-setup before the rest of the test can be recorded.

Note that depending on the video capture hardware and the device providing
the video stream, loss of video does not necessarily mean loss of frames
on the frame buffer.