Move default configuration to a location not writable by the user
The existing configuration search path was:

    1. /etc/stbt/stbt.conf
    2. ~/.config/stbt/stbt.conf
    3. $STBT_CONFIG_FILE

This commit adds a new path *before* the existing paths:

    0. /usr/libexec/stbt/stbt.conf
    1. ...
    2. ...
    3. ...

(assuming a typical installation prefix; exact paths may vary).

This is because /etc/stbt/stbt.conf is intended to be editable by the
user or the system administrator; if that user removes some
configuration entries, `stbt` will now continue to work.

Putting a config file in /usr/libexec may be a bit weird, but it is oh
so convenient for it to be in the same directory as stbt.py. According
to the Filesystem Hierarchy Standard (http://www.pathname.com/fhs/) it
should be in /usr/share, I suppose. On Fedora 17, packages like "rpm"
and "udev" store various configuration files (not executables, not
shared libraries) under /usr/lib/<package>/ so there is a precedent
(but that's lib, not libexec).

Note that `make install` installs stbt.conf to both
/usr/libexec/stbt/stbt.conf and /etc/stbt/stbt.conf. The latter serves
to document all available config options; the user can choose to remove
some entries, and `stbt` continues to work because it finds that entry
in /usr/libexec/stbt/stbt.conf.

Note that I didn't list /usr/libexec/stbt/stbt.conf in the README.rst /
man page, because it isn't something that users should be concerned with;
it's an implementation detail. /usr/lib* is for read-only data.