stbt.py: Cleanup: Move `get_config` function to the top of the file
This is needed by the next commit where I add a function that calls
`get_config` to set the function's default parameter values.

A function's default parameter values are evaluated when the function
is *defined*, not when the function is *called*; so `get_config` needs
to have been defined before that function.