`wait_for_match` etc. take `MatchParameters` argument instead of `**kwargs`
Commit dd44303b changed the `wait_for_match` signature from:

    wait_for_match(
        image,
        timeout_secs=10,
        consecutive_matches=1,
        match_method=1,
        match_threshold=0.8,
        confirm_method=2,
        erode_passes=1,
        confirm_threshold=None,
        noise_threshold=None)

to:

    wait_for_match(
        image,
        timeout_secs=10,
        consecutive_matches=1,
        noise_threshold=None,
        **kwargs)

because the default values for the templatematch-related parameters are
now loaded from stbt.conf.

That made the API slightly more obscure (yes it's documented in the man
page, but developers often like to refer to the source code directly),
and it also meant that tools like pylint won't catch mis-spellings of
those parameter names.

This commit makes the parameters explicit again, but instead of passing
them as individual keyword arguments to `wait_for_match`, you pass in
a `MatchParameters` object.

This saves duplicating all the image processing parameters as keyword
arguments in `wait_for_match`, `detect_match`, `press_until_match`, and
all user-defined functions that wrap the above (of which we have quite a
few at YouView, such as `wait_for_any_match`, `match`, and `match_any`).

The downside is that it makes the call sites quite a bit more verbose.
See for example this change made to one of the selftests in this commit:

    - wait_for_match("$scratchdir/gamut.png", confirm_threshold=0.7)
    + wait_for_match("$scratchdir/gamut.png",
    +                match_parameters=MatchParameters(confirm_threshold=0.7))

The argument in favour of this change is:

+ Advanced users of stb-tester will have quite a lot of their own
  user-defined functions that would benefit from this.

+ The uglyness at the call sites is something that shouldn't be very
  frequent; the image processing algorithm should *just work* most of
  the time without customisation (except for, possibly, global settings
  in stbt.conf).