Move documentation for `MatchParameters` into the source code
Note that this documentation still appears in README.rst, but now in the
section that is automatically populated from the source code docstrings.

A note regarding the changes to api-doc.sh:

We now want to print a class's full docstring, not just the first
paragraph (as delimited by a blank line). Before `MatchParameters`, none
of the classes' docstrings had more than one paragraph. So now we use
everything in the pydoc output up to "Methods defined here" (which pydoc
prints for classes with methods) or to "Method resolution order" (which
pydoc prints for all classes, and is the first thing to appear after the
docstring for classes with no methods explicitly defined in the source
code).