gsttemplatematch.c: GStreamer element documentation comment
I have no idea where the "#" and "%" notation comes from; I just copied
the notation from gstmotiondetect.c. The closest I found was by googling
for "gobject doc comments" which found this:
https://live.gnome.org/GObjectIntrospection/Annotations

which doesn't actually describe the "#" notation anywhere but does use
it in a couple of examples to refer to a type. Note though that the
above url refers to GObjectIntrospection which is fairly new; for
example it's used by GStreamer 1.0 but not by GStreamer 0.10 (the
version stb-tester uses).