stbt control: New tool to send remote control commands with PC keyboard
This tool is useful when multiple set-top-boxes are under test in the
same room, and it is not safe to use the remote control shipped with
the box to control one because the IR signals are received by multiple
set-top-boxes.

The tool uses the same remote control implementations as 'stbt run'.

A keymap file is required to save mappings of keyboard keys to remote
control keys. It was designed to make it easy to create one from scratch
manually. See `stbt remote --help` for detailed information.

It also has a mode to send a remote control key from the command line.