stbt config: API change from `[section] key` to `section.key`
This changes the command-line tool's API to:

    stbt config <section>.<key>

Since 9283df1, the search for a configuration key no longer falls back
to the "global" config section, so it makes less sense to provide a
default section. The new API is consistent with `git config`, and it
will allow us (in future) to add a parameter to specify the default
value should the key not be found in the configuration files.

I wanted to make this change now, before there are too many users of
`stbt config` (which was only added in the previous stb-tester release,
0.12). All current users of the tool will fail: Those specifying two
parameters will fail because now only one parameter is allowed; those
specifying a single parameter will fail because that parameter doesn't
include the section and the key name separated by a dot. So at least
there won't be any confusing silent changes in behaviour.