stbt.py: Deprecation messages should point at source location of caller
...not at the line in stbt.py with the call to `warnings.warn`.

Note that the default values for `stacklevel` is 1, and it is
interpreted as relative to the *implementation* of `warnings.warn`, not
its caller.

Changing the stacklevel to 2 makes the warning messages appear to belong
to the module of the *caller* of `wait_for_match` (etc.), not to the
`stbt` module, so with the existing `filterwarnings` specification the
warning messages wouldn't be printed at all. Hence the change to match
on the message contents instead of the module.

Unfortunately `stbt run` runs the script via `execfile`, so for scripts
executed directly via `stbt run`, the warning message will show "stbt"
for the source file (but the line number will the line number of the
offending statement in the actual script). I think that this commit is
still worthwhile because the source file will be printed correctly for
any calls to `wait_for_match` made from a *library* imported by the
top-level script. In our corpus of test scripts at YouView the vast
majority of calls to `wait_for_match` are made from such library
functions, so the source location printed by this change is of real
value.