stbt.py: Don't use object instance for `match_parameters` default value
A function's default parameter values are evaluated *once* when the
function is defined, not when the function is called; and the function
modifies this value, so subsequent calls to the same function could
get a nasty surprise if they're expecting the original default value.
For example:

(1)  def wait_for_match(..., match_parameters=MatchParameters()):
     ...

     stbt.wait_for_match("template.png", noise_threshold=0.1)
(2)  └ match_parameters.confirm_threshold = noise_threshold

     stbt.wait_for_match("template.png")
(3)  └ properties["confirmThreshold"] = match_parameters.confirm_threshold

All 3 of the numbered lines are referencing the same `MatchParameters`
instance, so the second call to `wait_for_match` sees the modification
that was made in the first call.

Thanks to Aiman Baharna for spotting this.