stbt templatematch: Backwards compatibility support for `noise_threshold`
In 18a5e70e `noise_threshold` was renamed to `confirm_threshold`.

In 25d7f3a9 `confirm_threshold` was moved into a field of
`MatchParameters`, and `stbt templatematch` was updated accordingly, but
in a way that lost backwards compatibility support for
`noise_threshold`.

We'll keep `noise_threshold` for a release or two, and then we'll remove
it.