stbt.get_config works from scripts run with `python`
(not just from scripts run with `stbt run`).

Before this commit, a python script calling `stbt.config` would give
different results if it was run with `python` or with `stbt run`. When
run directly with `python`, it wouldn't read the values from
`/etc/stbt/stbt.conf` (or the appropriate location for the host-wide
config file). This was because `stbt.get_config` was expecting the
environment variable `STBT_SYSTEM_CONFIG` to be set; this was set by
`stbt` (see `stbt.in`; the correct value for `STBT_SYSTEM_CONFIG` was
hard-coded into `stbt` when you'd run `make`).

This commit changes the mechanism for finding /etc/stbt/stbt.conf: Now
this value is hard-coded (again, at `make` time) into the default
read-only config file (e.g. /usr/libexec/stbt/stbt.conf -- see
1fc5016e).

In practice, any stb-tester script that uses the video-capture or
infrared control capabilities must be run with `stbt run`; but it's
conceivable for users to write helper scripts that only use
`get_config`.

A note on the `stbt.py` implementation: According to my experiments, you
can call `ConfigParser.read` multiple times without losing the
previously-loaded values, but this isn't documented. The
documentation[1] says to load defaults with `readfp` and then make a
single call to `read` passing in multiple filenames in an array.

[1] http://docs.python.org/2/library/configparser.html#ConfigParser.RawConfigParser.read

Fixes #17.