make check: Don't use PARALLEL environment variable
Commit 42d1d8aa uses PARALLEL if it is set in the environment, copying
the way that INSTALL and TAR are handled. This wasn't a good idea
because `parallel` uses `$PARALLEL` for its default options, so users
are likely to have PARALLEL set to something like "-j8".

Using `:=` instead of `?=` won't take the value from the environment,
but you can still override it with an explicit
`make check parallel=gparallel`.