make check: Run each test in a sub-shell to fix failures introduced by 9371c5d2
Commit 9371c5d2 had the test `test_wait_for_match_no_match` change the
working directory, so all subsequent tests failed when run sequentially.
(When run via GNU parallel, the problem didn't manifest because each
test is run in a different process.)

The solution is to always run each test in a sub-shell, so that a test
can freely modify its environment without affecting other tests.