test_detect_match_times_out_during_yield: Fix intermittent failure
and actually *test* something. Previously the invocation of `stbt-run`
would always succeed (though it was occasionally failing on heavily
loaded systems because `timeout` kicked in before `detect_match` timed
out; now I've removed that timeout because it is completely
unnecessary).

Now the test asserts that `detect_match` only yields a single value,
(the timeout specified in `timeout_secs` should cause the `for` loop to
end after the first iteration).