test_wait_for_match_nonexistent_template: Fix intermittent failure
Seen on our Travis continuous integration service:
https://travis-ci.org/drothlis/stb-tester/builds/7600784

Presumably on a heavily-loaded system, the test timed out before
`stbt-run` had a chance to run the `wait_for_match` function.

This test doesn't need a timeout at all, because (if `stbt-run` is
working correctly) `stbt-run` will always terminate.  If `stbt-run`
is so buggy that it hangs, someone will notice that the test hangs!