The unit test `test_match_consecutive_timed_frames` was setting the frame rate of the output, which does not affect the test result. So, in this context, that is useless and  has been removed.
The unit test `test_live_stream_caught_up_after_process_all_frames` must
check that after leaving the "process all frames" context, it is
displyaing the expected output. This test case has nothing to do with
accuracy, so in order to avoid confusion, the timed frames are not
referenced in it.