stbt._log_image: Scale "matchTemplate" image to visible range
The result of "cv2.matchTemplate" is a matrix with each pixel in the
range [0.0, 1.0]. This change scales the debug image saved to disk
to the grayscale range [0, 255].

Note that "_log_image" reassigns the name "image" to the scaled version,
but this name is local to the function. The scaling isn't done in-place,
so the original image isn't modified (otherwise the
"first_pass_certainty" comparison in "_match_template" would fail).