stbt.py: Run GLib main loop continuously, and wait on appsink "new-buffer" signals
...instead of polling appsink's "last-buffer" property.

Now we will monitor pipeline signals & messages like "underrun" all the
time, not just when inside a call to "wait_for_match" (or "detect_match",
"wait_for_motion", etc). So, for example, we will restart the pipeline
when we lose video input (see b6135eca) as soon as we detect video loss,
not just at the next "wait_for_match".

The GLib main loop is now running in a separate thread (not in the main
application thread, which runs the user's script), so "sys.exit" doesn't
work (it only raises a SystemExit exception, which terminates the thread
but isn't propagated to the main application thread). So we use the same
synchronised Queue that we use to communicate new frames to the main
application thread, for also communicating errors (in the form of
Exception objects).