stbt control: Do not raise `curses.error` if the terminal is too small
`stbt control` uses `curses` module for terminal handling. `curses`
raises the unhelpful `curses.error` exception if the terminal is smaller
than what is required to print the keymap.

If the terminal is too small do not print the keymap and print an error
message instead, but do not quit the tool as the more experienced users
who remember key mappings might not need the printed keymap at all.