Disable Travis CI (remove .travis.yml config file)
travis-ci.org is a free continuous-integration service. I've disabled
it because "test_wait_for_motion_with_mask_does_not_report_motion"
fails, possibly due to a bug in OpenCV 2.3:
http://answers.opencv.org/question/15279/

And we run "make check" ourselves during development anyway. We do lose
the opportunity to test on Ubuntu; but if "make check" always fails on
travis-ci, it has no value.

Travis-ci uses Ubuntu 12.04 LTS, which has OpenCV 2.3. I tried to
install OpenCV 2.4 from this PPA:
https://launchpad.net/~limparissoft/+archive/opencv-ppa

by applying the following patch to .travis.yml:

    diff --git a/.travis.yml b/.travis.yml
    index b00a049..3c481e6 100644
    --- a/.travis.yml
    +++ b/.travis.yml
    @@ -1,6 +1,7 @@
     language: c

     install:
    + - sudo add-apt-repository ppa:limparissoft/opencv-ppa
      - sudo apt-get update
      - sudo apt-get install gstreamer0.10-plugins-bad gstreamer0.10-ffmpeg
                             libgstreamer0.10-dev python-gst0.10 python-docutils
    @@ -8,7 +9,7 @@ install:
                             libcv-dev libhighgui-dev python-nose xvfb
                             gstreamer0.10-tools gstreamer0.10-x
                             gstreamer0.10-plugins-good pylint pep8
    -                        python-numpy python-opencv
    +                        opencv=2.4.5-0~3~precise1 python-numpy python-opencv
      - make

     script:

but it fails to install due to a conflict with opencv 2.3 (installed as
a dependency of python-opencv; the PPA doesn't have a 2.4 package for
python-opencv).