tests: Two new tests have been added for the new feature `process_all_frames`.
When using this context, all the frames must be processed and there must
not be any frame loss. In order to test this, the framerate is forced to be
50fps, and afterwards the timestamp of the frames is overlapped. That way,
the expected overlapped strings are: "00:00:00.000", "00:00:00.020",
"00:00:00.040" and so on.

Due to a dependency on the speed of the computer where this is run, the
first match must not be at any specific time. Otherwise, it can be
missed and therefore, fail. Instead, the milliseconds are checked, and
afterwards, the timestamps will report whether the matches were
consecutive.