stbt.detect_motion: Fix invalid use of mask in OpenCV operation
The "mask" parameter to "bitwise_and" means that the operation only
changes those destination array pixels "p" for which "mask(p)" != 0.
Since we're not specifying "dst" (destination), "bitwise_and" allocates
a new output array, but doesn't initialize it properly, so anything in
the output image in the area corresponding to the 0 pixels in the mask
contain garbage, rather than 0. See
http://code.opencv.org/issues/1748#note-2

This fixes the behaviour described in
http://answers.opencv.org/question/15279

Presumably the reason I didn't see this behaviour in OpenCV 2.4.5 is
because of http://code.opencv.org/issues/1286