stbt.py: Send End-Of-Stream before exiting
And wait for the EOS be fully processed by all sink elements, to allow
any muxers and filesinks specified by the user in `--sink-pipeline` to
write any index or header information. Without this, video recordings
of the test run[1] might be corrupted or not be seekable.

We can't send the EOS event to the pipeline as a whole because some
source elements don't propagate the EOS event properly[2]. So we send
the EOS to the sink bin's sink pad instead[3]. Each sink element in the
sink bin will post an EOS *message* after the element has fully
processed the EOS event, and the sink bin will post an EOS message to
the pipeline once it has received EOS messages from *all* its sink
elements. The pipeline won't send an EOS message to the application
until all the bins have posted EOS messages[4], so we go ahead and post
an EOS message on the source bin's behalf.

The final `pipeline.set_state(gst.STATE_NULL)` doesn't seem to be
necessary, and in fact it prevented the EOS message from reaching the
application. After removing it I've tested with videotestsrc,
v4l2src (with the Hauppauge HDPVR) and decklinksrc, and I don't see any
harmful effects. That line had existed since the initial commit, in
`stbt-run`.

Note that if teardown is called in the middle of `restart_source_bin`,
`stbt-run` hangs because the sink bin fails to process the EOS event,
so the GLib mainloop is never quit. The log from `GST_DEBUG=*:6` says:

    Received event on flushing pad. Discarding

...referring to the EOS event received on the sink bin's sink pad, which
is in the "flushing" state.

The unit test requires `webminspector.py` from
http://git.chromium.org/gitweb/?p=webm/webminspector.git
with this patch (pending review):
https://gerrit.chromium.org/gerrit/#/c/59912/
If `webminspector.py` is not found, the test won't be run.

On OS X the recorded video is always invalid (perhaps a bug with the
vp8enc or webmmux elements), so don't try to run the unit test on OS X.

[1] e.g. with a sink-pipeline of:

        vp8enc speed=7 ! webmmux ! filesink location=video.webm

[2] In particularly decklinksrc, the GStreamer source element for
    Blackmagic Intensity Pro video-capture cards. This is fixed on the
    GStreamer 1.0 branch:
    https://bugzilla.gnome.org/show_bug.cgi?id=680700
    but stb-tester still uses GStreamer 0.10.

[3] This post on the mailing list was helpful:
    http://gstreamer-devel.966125.n4.nabble.com/Handling-EOS-in-a-branched-pipeline-tp2965563.html

[4] In debugging this I used `GST_DEBUG=*:6` and grepped for `eos` as
    suggested here:
    http://gstreamer-devel.966125.n4.nabble.com/Pipeline-Not-Shut-Down-After-End-of-Stream-tp4659955p4659957.html