stbt.get_frame: Wait for the next available frame
If called before any frames are available (e.g. before the GStreamer
pipeline has fully initialised, or if the system-under-test isn't
delivering any video), `get_frame` will wait up to `timeout_secs`
seconds. Previously `get_frame` returned `None` if no video was
available.

One minor consequence of this is that the screenshot attached to the
`MatchTimeout` raised by `wait_for_match` is *guaranteed* to be a frame
*after* the last frame analysed by `wait_for_match`, instead of
only *sometimes* being a different frame. Maybe we can attach the actual
frame to each `MatchResult` object, but that's for a different commit.

Fixes #23, fixes #33.