stbt.py: Wait for GLib main loop to finish before exiting
This is a cleanup commit that shouldn't change the observable behaviour.
It is necessary for the next commit, where we wait for end-of-stream
events to be processed by all sinks in the pipeline before we exit.

Note that `on_error` will be called at most once, after which
`on_new_buffer` shouldn't ever be called because we have quit the GLib
main loop, so the checks in `tell_user_thread` are no longer necessary.