stbt.py: Trigger source restarts from underruns in the *application* pipeline
Not from underruns in the *source* pipeline. This should catch any
problems in the communication between the source and application
pipelines, not just problems within the source pipeline itself.

I don't think I've seen any actual problem that this solves, but it
seems more robust this way, and the code reflects more closely
the *goal* of the source restarting behaviour.