stbt.py: Source-restart-on-video-loss behaviour is configurable
Commit 5a0770e6 made this behaviour configurable *and* also changed the
way source restarting is done. 5a0770e6 was reverted in a189e0c9 because
the new way of restarting the source was unreliable. This current commit
applies the configurability changes from 5a0770e6 without changing the
source restarting mechanism.

We want this to be configurable because some video sources don't *need*
this behaviour. Only the Hauppauge HD PVR needs it. With the Blackmagic,
for example, we prefer to block until video is re-established (or until
"stbt.Display.get_frame" times out) because it is simpler, the logs are
less polluted with "stopping"/"starting" messages, and if you're
recording a video with "--save-video" the video only contains what
happened since the last restart.

Note that the config value accepts "1", "yes", "true", or "on" (case
insensitively), unlike 5a0770e6 which only accepted "True". This is
consistent with ConfigParser.getboolean -- unfortunately our
"get_config" wrapper around ConfigParser doesn't allow us to call
getboolean directly.