stbt.Display.teardown: Avoid intermittent error from decklinksrc
Occasionally, on teardown the following GStreamer error would happen
*before* the pipeline received the EOS event, so if a video was being
recorded it wouldn't have the index information necessary for seeking.

    Internal GStreamer error: negotiation problem.
    gstdecklinksrc.cpp(1248): gst_decklink_src_task ():
      /GstPipeline:stb-tester/GstBin:bin0/GstDecklinkSrc:decklinksrc0

It is still necessary to post an EOS message to the source bin, even
after unlinking it from the sink bin (see b7a37493).

I tested this with:

    while ./stbt-run --save-video video.webm -vv test.py &&
          webminspector.py video.webm | grep 'Cue Point';
    do
        :
    done

which would fail after 3 or 4 runs before this commit, and succeeded 50
times with this commit. See b7a37493 for details on webminspector.py.